Split vagrant setup into a base image and developer environment
Some of the FOSS components required for our development environment
require local compilation or otherwise expensive operations.  This
impacts the time to "vagrant up" the development environment, which
can be problematic on slower machines and/or in cases where frequent
rebuilds are the norm.

This patch injects an intermediate step in the form of creating a
base vagrant "box" (in this case "openchain/devbase") to supplant
the generic "ubuntu/trust64" box currently in use.  The idea is that
the upstream components may be incorporated into the base box at
a frequency that is expected to be much lower than a normal developer
environment lifecycle.

Usage: One could describe the current workflow as

1) vagrant box add ubuntu/trusty64 (when required)
2) vagrant up (wait 20+ minutes)
3) vagrant ssh
4) vagrant destroy
5) go to step 2

With this patch, there is an intermediate step to construct our custom
box

1) vagrant box add ubuntu (when required)
2) cd baseimage; make install (wait 20+ minutes)
3) vagrant up (quick)
4) vagrant ssh
5) vagrant destroy
6) go to step 3

While the loop between 3-6 is a substantial improvement over the current
system, (2) still imposes an undesirable overhead for new installations.

The plan/proposal is that if this patch and its methodology are accepted,
the resulting image may be published to a service such as
atlas.hashicorp.com.  Were that to happen, steps (1) and (2) could be
collapsed down to "vagrant box add openchain/devbase", which should be
significantly faster in most cases unless the user is behind a
particularly slow internet connection.  In those cases, the user may
opt to build the image manually as per above.