Replace misleading (expected) failure message during baseimage creation
We want to optimize the baseimage filesystem to be as small as possible.  To achieve this, we
employ a strategy of writing all zeros to a file (/EMPTY) until the filesystem runs outs of
space, and then delete the file.  This allows the subsequent compression of the filesystem
to optimize away any of the unused portions.  However, dd will report an error once the
filesystem runs out of space.  This error can be confusing to someone auditing the baseimage build.

Therefore, this patch recognizes that the dd failure is uninteresting, and buries it in /dev/null
while simultaneously printing a user friendly reminder of whats occurring.

Signed-off-by: Gregory Haskins <gregory.haskins@gmail.com>